+++ packages/taro-ui/rn/common/utils.ts
@@ -17,14 +17,16 @@ function delayQuerySelector(
   delayTime = 500
 ): Promise<any[]> {
   return new Promise(resolve => {
-    const selector: SelectorQuery = Taro.createSelectorQuery()
-    delay(delayTime).then(() => {
-      selector
-        .select(selectorStr)
-        .boundingClientRect()
-        .exec((res: any[]) => {
-          resolve(res)
-        })
+    Taro.nextTick(() => {
+      const selector: SelectorQuery = Taro.createSelectorQuery()
+      delay(delayTime).then(() => {
+        selector
+          .select(selectorStr)
+          .boundingClientRect()
+          .exec((res: any[]) => {
+            resolve(res)
+          })
+      })
     })
   })
 }
@@ -212,13 +214,8 @@ function handleTouchScroll(flag: any): void {
 
 function pxTransform(size: number): string {
   if (!size) return ''
-  const designWidth = 750
-  const deviceRatio = {
-    640: 2.34 / 2,
-    750: 1,
-    828: 1.81 / 2
-  }
-  return `${size / deviceRatio[designWidth]}rpx`
+
+  return Taro.pxTransform(size)
 }
 
 function objectToString(style: object | string): string {