+++ packages/taro-ui/src/components/image-picker/index.tsx
@@ -75,7 +75,9 @@ export default class AtImagePicker extends React.Component<AtImagePickerProps> {
         const newFiles = files.concat(targetFiles)
         this.props.onChange(newFiles, 'add')
       })
-      .catch(this.props.onFail)
+      .catch(err => {
+        this.props?.onFail?.(err)
+      })
   }
 
   private handleImageClick = (idx: number): void => {
@@ -134,17 +136,18 @@ export default class AtImagePicker extends React.Component<AtImagePickerProps> {
               ) : (
                 <View
                   className='at-image-picker__flex-item'
-                  key={'empty_' + i * length + j}
+                  key={`empty_${i * length}${j}`}
                 >
                   {item.type === 'btn' && (
-                    <View
-                      className='at-image-picker__item at-image-picker__choose-btn'
-                      onClick={this.chooseFile}
-                    >
-                      <View className='add-bar'></View>
-                      <View className='add-bar'></View>
-                    </View>
-                  )}
+                    <View onClick={this.chooseFile}>
+                      {
+                        this.props.children
+                        || <View className='at-image-picker__item at-image-picker__choose-btn'>
+                          <View className='add-bar'></View>
+                          <View className='add-bar'></View>
+                        </View>
+                      }
+                    </View>)}
                 </View>
               )
             )}