+++ packages/taro-ui/src/components/list/item/index.tsx
@@ -40,13 +40,11 @@ export default class AtListItem extends React.Component<AtListItemProps> {
       hasBorder,
       extraThumb,
       switchColor,
-      switchIsCheck
+      switchIsCheck,
+      icon
     } = this.props
 
-    let { extraText, title } = this.props
-
-    extraText = String(extraText)
-    title = String(title)
+    const { extraText, title } = this.props
 
     const rootClass = classNames(
       'at-list__item',
@@ -58,42 +56,51 @@ export default class AtListItem extends React.Component<AtListItemProps> {
       },
       this.props.className
     )
-    const iconClass = classNames(
-      (iconInfo && iconInfo.prefixClass) || 'at-icon',
-      {
-        [`${(iconInfo && iconInfo.prefixClass) || 'at-icon'}-${
-          iconInfo && iconInfo.value
-        }`]: iconInfo && iconInfo.value
-      },
-      iconInfo && iconInfo.className
-    )
+
+    const renderIcon = () => {
+      if (icon) {
+        return <View className='item-icon'>{icon}</View>
+      } else if (iconInfo?.value) {
+        const iconClass = classNames(
+          (iconInfo && iconInfo.prefixClass) || 'at-icon',
+          {
+            [`${(iconInfo && iconInfo.prefixClass) || 'at-icon'}-${
+              iconInfo && iconInfo.value
+            }`]: iconInfo && iconInfo.value
+          },
+          iconInfo && iconInfo.className
+        )
+        return (
+          <View className='at-list__item-icon item-icon'>
+            <Text
+              className={iconClass}
+              style={mergeStyle(
+                {
+                  color: iconInfo.color || '',
+                  fontSize: `${iconInfo.size || 24}px`
+                },
+                iconInfo.customStyle || ''
+              )}
+            ></Text>
+          </View>
+        )
+      } else if (thumb) {
+        return (
+          <View className='at-list__item-thumb item-thumb'>
+            <Image
+              className='item-thumb__info'
+              mode='scaleToFill'
+              src={thumb}
+            />
+          </View>
+        )
+      }
+    }
 
     return (
       <View className={rootClass} onClick={this.handleClick}>
         <View className='at-list__item-container'>
-          {thumb && (
-            <View className='at-list__item-thumb item-thumb'>
-              <Image
-                className='item-thumb__info'
-                mode='scaleToFill'
-                src={thumb}
-              />
-            </View>
-          )}
-          {iconInfo && iconInfo.value && (
-            <View className='at-list__item-icon item-icon'>
-              <Text
-                className={iconClass}
-                style={mergeStyle(
-                  {
-                    color: iconInfo.color || '',
-                    fontSize: `${iconInfo.size || 24}px`
-                  },
-                  iconInfo.customStyle || ''
-                )}
-              ></Text>
-            </View>
-          )}
+          {renderIcon()}
           <View className='at-list__item-content item-content'>
             <View className='item-content__info'>
               <View className='item-content__info-title'>{title}</View>
@@ -156,16 +163,16 @@ AtListItem.defaultProps = {
 }
 
 AtListItem.propTypes = {
-  note: PropTypes.string,
+  note: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
   disabled: PropTypes.bool,
-  title: PropTypes.string,
+  title: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
   thumb: PropTypes.string,
   onClick: PropTypes.func,
   isSwitch: PropTypes.bool,
   hasBorder: PropTypes.bool,
   switchColor: PropTypes.string,
   switchIsCheck: PropTypes.bool,
-  extraText: PropTypes.string,
+  extraText: PropTypes.oneOfType([PropTypes.string, PropTypes.element]),
   extraThumb: PropTypes.string,
   onSwitchChange: PropTypes.func,
   arrow: PropTypes.oneOf(['up', 'down', 'right']),